MINIFICPP-659: Break out CAPI into nanofi
Break out structure -- facilitating Python and new nanofi structure
