+++ cmake/BuildTests.cmake
@@ -29,12 +29,13 @@ MACRO(GETSOURCEFILES result curdir)
   SET(${result} ${dirlist})
 ENDMACRO()
 
+set(NANOFI_TEST_DIR "${CMAKE_SOURCE_DIR}/nanofi/tests/")
+
 if(NOT EXCLUDE_BOOST)
 	find_package(Boost COMPONENTS system filesystem)
 endif()
 
-function(createTests testName)
-   message ("-- Adding test: ${testName}")
+function(appendIncludes testName)
     target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/thirdparty/catch")
     target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/thirdparty/spdlog-20170710/include")
     target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/thirdparty/yaml-cpp-yaml-cpp-0.5.3/include")
@@ -60,6 +61,11 @@ function(createTests testName)
     target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/libminifi/include/utils")
     target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/libminifi/include/processors")
     target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/libminifi/include/provenance")
+endfunction()
+
+function(createTests testName)
+    message ("-- Adding test: ${testName}")
+    appendIncludes("${testName}")
 
 	if (ENABLE_BINARY_DIFF)
     	target_include_directories(${testName} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/thirdparty/bsdiff/")
@@ -70,11 +76,13 @@ function(createTests testName)
     endif()
     target_link_libraries(${testName} ${CMAKE_DL_LIBS} ${SPD_LIB} ${TEST_BASE_LIB})
     target_link_libraries(${testName} ${CMAKE_THREAD_LIBS_INIT} ${OPENSSL_LIBRARIES} core-minifi yaml-cpp)
-    if (APPLE)
-		target_link_libraries (${testName} -Wl,-all_load minifi)
-	else ()
-		target_link_libraries (${testName} -Wl,--whole-archive minifi -Wl,--no-whole-archive)
-	endif ()
+    if (NOT excludeBase)
+      if (APPLE)
+  		target_link_libraries (${testName} -Wl,-all_load minifi)
+	  else ()
+  		target_link_libraries (${testName} -Wl,--whole-archive minifi -Wl,--no-whole-archive)
+	  endif ()
+	endif()
     if (Boost_FOUND)
         target_link_libraries(${testName} ${Boost_SYSTEM_LIBRARY})
         target_link_libraries(${testName} ${Boost_FILESYSTEM_LIBRARY})
@@ -105,7 +113,7 @@ target_include_directories(${CATCH_MAIN_LIB} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}
 SET(TEST_RESOURCES ${TEST_DIR}/resources)
 
 GETSOURCEFILES(UNIT_TESTS "${TEST_DIR}/unit/")
-GETSOURCEFILES(CAPI_UNIT_TESTS "${TEST_DIR}/capi/")
+GETSOURCEFILES(NANOFI_UNIT_TESTS "${NANOFI_TEST_DIR}")
 GETSOURCEFILES(INTEGRATION_TESTS "${TEST_DIR}/integration/")
 
 SET(UNIT_TEST_COUNT 0)
@@ -120,15 +128,17 @@ ENDFOREACH()
 message("-- Finished building ${UNIT_TEST_COUNT} unit test file(s)...")
 
 SET(UNIT_TEST_COUNT 0)
-FOREACH(testfile ${CAPI_UNIT_TESTS})
+FOREACH(testfile ${NANOFI_UNIT_TESTS})
     get_filename_component(testfilename "${testfile}" NAME_WE)
-    add_executable("${testfilename}" "${TEST_DIR}/capi/${testfile}")
-    createTests("${testfilename}")
-    target_link_libraries(${testfilename} ${CATCH_MAIN_LIB} capi)
+    add_executable("${testfilename}" "${NANOFI_TEST_DIR}/${testfile}")
+    target_include_directories(${testfilename} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/nanofi/include")
+    target_include_directories(${testfilename} BEFORE PRIVATE "${CMAKE_SOURCE_DIR}/libminifi/test")
+    appendIncludes("${testfilename}")
+    target_link_libraries(${testfilename} ${CMAKE_THREAD_LIBS_INIT} ${CATCH_MAIN_LIB} nanofi)
     MATH(EXPR UNIT_TEST_COUNT "${UNIT_TEST_COUNT}+1")
     add_test(NAME "${testfilename}" COMMAND "${testfilename}" WORKING_DIRECTORY ${TEST_DIR})
 ENDFOREACH()
-message("-- Finished building ${UNIT_TEST_COUNT} capi unit test file(s)...")
+message("-- Finished building ${UNIT_TEST_COUNT} NanoFi unit test file(s)...")
 
 SET(INT_TEST_COUNT 0)
 FOREACH(testfile ${INTEGRATION_TESTS})
