+++ nanofi/examples/generate_flow.c
@@ -20,74 +20,44 @@
 #include <string.h>
 #include <sys/types.h>
 #include <sys/stat.h>
-#include <unistd.h>
-#include <dirent.h>
-#include <pthread.h>
 
-#include "file_blocks.h"
-#include "comms.h"
-#include "capi/api.h"
-#include "capi/processors.h"
-int is_dir(const char *path) {
-  struct stat stat_struct;
-  if (stat(path, &stat_struct) != 0)
-    return 0;
-  return S_ISDIR(stat_struct.st_mode);
-}
-
-pthread_mutex_t mutex;
-pthread_cond_t condition;
-
-int stopped;
-
-int stop_callback(char *c) {
-  pthread_mutex_lock(&mutex);
-  stopped = 1;
-  pthread_cond_signal(&condition);
-  pthread_mutex_unlock(&mutex);
-  return 0;
-}
-
-int is_stopped(void *ptr) {
-  int is_stop = 0;
-  pthread_mutex_lock(&mutex);
-  is_stop = stopped;
-  pthread_mutex_unlock(&mutex);
-  return is_stop;
-}
+#include "api/nanofi.h"
 
 /**
  * This is an example of the C API that transmits a flow file to a remote instance.
  */
 int main(int argc, char **argv) {
-  if (argc < 5) {
-    printf("Error: must run ./monitor_directory <instance> <remote port> <directory to monitor>\n");
+
+  if (argc < 3) {
+    printf("Error: must run ./generate_flow <instance> <remote port> \n");
     exit(1);
   }
 
-  stopped = 0x00;
-
   char *instance_str = argv[1];
   char *portStr = argv[2];
-  char *directory = argv[3];
 
   nifi_port port;
 
   port.port_id = portStr;
 
-  C2_Server server;
-  server.url = argv[4];
-  server.ack_url = argv[5];
-  server.identifier = "monitor_directory";
-  server.type = REST;
-
   nifi_instance *instance = create_instance(instance_str, &port);
 
-  // enable_async_c2(instance, &server, stop_callback, NULL, NULL);
+  flow *new_flow = create_flow(instance, "GenerateFlowFile");
 
-  flow *new_flow = monitor_directory(instance, directory, 0x00, KEEP_SOURCE);
+  flow_file_record *record = get_next_flow_file(instance, new_flow);
 
-  transmit_to_nifi(instance, new_flow, is_stopped);
+  if (record == 0) {
+    printf("Could not create flow file");
+    exit(1);
+  }
+
+  transmit_flowfile(record, instance);
+
+  free_flowfile(record);
+
+  // initializing will make the transmission slightly more efficient.
+  //initialize_instance(instance);
+  //transfer_file_or_directory(instance,file);
 
   free_flow(new_flow);
